Google Maps directions intent and refactoring changes
