+++ app/build.gradle
@@ -35,12 +35,12 @@ dependencies {
     implementation 'com.android.support.constraint:constraint-layout:1.1.3'
     implementation "android.arch.lifecycle:viewmodel:1.1.1"
     implementation 'android.arch.lifecycle:extensions:1.1.1'
-    implementation 'com.dinuscxj:recyclerrefreshlayout:2.0.5'
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
 
     implementation 'com.android.support:design:28.0.0'
+    implementation 'com.dinuscxj:recyclerrefreshlayout:2.0.5'
     implementation 'com.google.android.gms:play-services-maps:16.0.0'
     implementation 'com.google.zxing:core:3.3.0'
     implementation 'com.squareup.retrofit2:converter-gson:2.4.0'
