+++ app/src/main/java/org/hackillinois/android/common/TimeInfo.kt
@@ -1,4 +1,4 @@
-package org.hackillinois.android.utils
+package org.hackillinois.android.common
 
 class TimeInfo(millis: Long) {
     val days: Long
