+++ app/src/main/java/org/hackillinois/android/view/EventInfoActivity.kt
@@ -5,8 +5,8 @@ import android.support.v7.app.AppCompatActivity
 import kotlinx.android.synthetic.main.activity_event_info.*
 import org.hackillinois.android.R
 import org.hackillinois.android.model.Event
-import android.content.Intent
-import android.net.Uri
+import com.google.android.gms.maps.model.LatLng
+import org.hackillinois.android.common.DirectionsOnClickListener
 
 class EventInfoActivity : AppCompatActivity() {
 
@@ -20,10 +20,7 @@ class EventInfoActivity : AppCompatActivity() {
         eventLocation.text = event.locationDescription
         eventDescription.text = event.description
 
-        directionsButton.setOnClickListener {
-            val geoUri = "http://maps.google.com/maps?q=loc:${event.latitude},${event.longitude} (${event.locationDescription})"
-            val intent = Intent(Intent.ACTION_VIEW, Uri.parse(geoUri))
-            startActivity(intent)
-        }
+        val location = LatLng(event.latitude, event.longitude)
+        directionsButton.setOnClickListener(DirectionsOnClickListener(location, event.name))
     }
 }
